Changed the format of URLs used by the fits package:
To refer to the primary URL of a FITS file you used to say 'file.fits'
or 'file.fits#1', and to the first extension (second HDU) you used to
say 'file.fits#2'

Now you refer to the primary HDU as 'file.fits' or 'file.fits#0'
and to the first extension as 'file.fits#1'.

I've changed it to be compatible with nom.tam.fits and the VOTable
FITS extnum attribute.

Note this will break existing NDXs stored as FITS files or as XML files
which reference FITS files.