BridgeNdx now calls HdxResourceType.newHdxResourceType to create and initialise the ndx, image, variance and quality Hdx types.
Added getDOMFacade and protected non-static class BridgeNdxDOMFacade,
which implements uk.ac.starlink.hdx.DOMFacade.  toXML method
repurposed into protected method constructDOM, and replaced by a toXML
which consists of { return getDOMFacade().getSource(); }.  Since this
duplicates functionality, I've marked toXML as deprecated, but not
removed it.  It should be easy to have MutableNdx extend
BridgeNdxDOMFacade and support the DOM-mutation elements, too.

ConstructDOM (was toXML) uses HdxDOMImplementation's HdxDocument to
create its DOM, which means it is able to incorporate into the DOM the
DOMFacade instances returned by get{Image,Variance,Quality}().  Also,
the URIs are now included in the DOM as the value of "url" attributes
rather than element content: we need to discuss this -- I think it
makes things simpler and loses nothing if there is no text content in
elements, only attributes -- and make corresponding changes
elsewhere.  Because of this difference, there's a difference between
the way that NdxIO writes Ndxes as XML and the way that constructDOM's
DOM is serialised, which causes regression tests to fail when there's
nothing actually wrong.  The original version of constructDOM is still
there, commented, out, in case I've done too much and it's necessary
to revert.

Added method getHdxType to return the HdxResourceType corresponding to
Ndx