Modified the way that treeview invokes SPLAT and SoG. It still talks using SOAP to instances of the SPLAT/SoG viewer services if they exist at their well-known endpoints.  Behaviour if the services don't exist is changed however. It used to try to start these services up in a new JVM, which was non-java and hence non-portable and also somewhat unreliable.  Now it instantiates suitable viewer classes in its own JVM in this case.
Related to this, a new standalone application (jar file plus script)
called showndx is introduced.  This just invokes an appropriate viewer
for a given NDX (depending on dimensionality).  This follows a suggestion
of Andy Adamson (JACH) who wanted a single viewer app he could associate
with .sdf files in the Windows (or CDE etc) file browser.