jni: Rebuild the OSX 64-bit libraries so that they use -install_name
The JNI libraries on OSX were built with the build path burned in as
the install name. This was preventing SPLAT from loading its JNI
library because it depends on libjniast. At build time libsplat was
configured to load libjniast from the jniast build directory and
for all except the person building SPLAT this directory was not
present.

Now force -install_name of JNI libraries to be the actual
installed name (/star/starjava/lib/$arch/libx.jnilib). SPLAT
now seems to load properly.