util: add gitversion utility for stamping standalone jars.
gitversion provides a replacement for svnversion now the source code
is managed by git.  The git show/describe commands don't seem to do
quite what is required, so this is a short shell script that does.
It is used to replace svnversion in the ttools and topcat
build-standalone targets to place fixed version strings into the
published jar files which can report build versions at runtime.