topcat: better display of aux column metadata.
Non-standard per-column "auxiliary" metadata is now displayed better,
in particular if the aux metadata type is array-valued the array
elements are shown rather than just calling toString() on the array object.
It's not perfect though; if the aux metadata type is Object (because
different columns have values of different types) it will still revert to
the old behaviour.  This case is unusual, but not unheard of.