Fixed multi-cone search to handle interruptions carefully.
The ConeErrorPolicy class interface and implementations have been modified
so that any time the multi-cone thread is interrupted, it makes sure that
no more cone search queries are presented.  Previously, although
interruptions were checked for, it was only done when a successful
result was obtained, rather than when a query was made.  Especially
in the case that a service was failing (e.g. due to overload) that
could mean that the client kept on hammering the service with queries
whose results would never get looked at, until JVM shutdown.