topcat: ensure that ConeMatcher can be cancelled properly.
Modify the ConeMatcher class returns an object that can be run and
interrupted rather than returning a TableProducer with an
unreferenced thread that the caller cannot access.  This should fix
it so that hitting Stop in the multi-cone window really does shut
down all threads associated with the multi-cone sequence.
However, the STILTS coneskymatch could still in principle leave
threads running if the task is interrupted but the JVM is not shut down.