votable: treat XML parse errors as VOTable format errors.
java.io.CharConversionErrors (which extend IOException) when parsing
a VOTable were getting thrown as IOExceptions, hence scuppering the
entire table load attempt.  Fixed it so they get thrown as
TableFormatExceptions, which in auto mode are cauhgt and allow attempts
to load the table in other (non-XML) formats.  Did the same thing for
SAXParseExceptions as well.