ttools: turn off precision squashing in layer plots.
There is a switch that stores cached double-precision column values as
single precision in the plot2 framework.  This was turned on, since
for most purposes the result makes no visible difference and it cuts
the required cache storage size in two.  However, it's dangerous,
and in some cases disastrous (e.g. for CDF EPOCH data, which are
milliseconds since AD0, values now ~6e13, it results in quantising
time values to the order of about an hour).  This is still a promising
optimisation, but if I use it I'll have to make sure that it's safe
to do so for the plot at hand.  This will require some effort.