topcat: work around java 1.5 xerces bug using build file properties.
There is a xerces bug specific to java 1.5 that intermittently causes
the XML processing that builds the user document to fail.  It's fixed
at java 1.6, but we don't really want to do the whole build for release
using 1.6 because we want 1.5-compatible class files.  So provide the
option to use an alternative JVM when doing the document construction
parts of the build.  This can be used by setting the alt.java
property in ~/.stardev.properties.  Anybody using 1.6+ already to
do the build won't need to do this.

An alternative approach would be to cross-compile - this would need
changes to all the starjava build.xml files to add suitable options to
the <javac> task invocations.