topcat: Add missing plot2 change log information.
Add summary of new plotting features introduced in v4.0b in the
v4.0b entry of the history in sun253.  This was missing from the
docs in the v4.0b release due to an oversight, but better to edit
it for later reference than to leave it out for ever.