table: Remove JTable presentation items from ValueInfo.
ValueInfo used to contain getTableCellRenderer, and DefaultValueInfo
had getTableCellEditor as well.  These really didn't belong there
and as far as I know were not used outside the table package, so
move that into uk.ac.starlink.table.gui(.StarTableColumn) where
they are happier.  I may provide another interface later for tables
that want to provide customised JTable GUI options.

This is a change to a class that's fundamenatal to the STIL API -
but probably not a big deal even so because I doubt anybody used
these methods.