ttools: Modify the way that DataGeom works with PlotLayers.
DataGeom is now OK to use if it only has some coordinates, i.e. it
might specify a point, or a line, or a plane or whatever.
It has a new hasPosition() method to report which of these it's doing.
This means that non-point-plotting layers (such as spectrogram) can use
the DataGeom machinery to report ranges.  Users of DataGeom have to
be a bit more careful now.

Rename Plotter.getDataGeoms to getPointDataGeoms, to emphasise
that these ones must be full-point (hasPosition=true) instances.

Also add PlotLayer.extendCoordinateRanges which gives you the chance to
report non-point-like range extensions (like a spectrogram channel array).