topcat: add new Highlight Rows action.
When a new subset is created, its rows are selected in the table viewer
window and it is made visible in all plots.  However, as pointed out
by Karl Glazebrook, there is no easy way to force this action for an
existing subset.  Added an action to do this in the Subsets window,
and a similar action in the table viewer window which just highlights the
rows without affecting plots.  Also withdrew the (slightly dangerous,
undocumented, and iconless) Apply Subset action in the table viewer window.