table: Add DomainMapper.
A domain mapper allows the possibility to associate conversion behaviour
with a ValueInfo (e.g. a column).  It's necessary to have something like
this in order for the CDF table reader to be able to communicate information
downstream about how to make sense of various CDF-specific time encodings
(CDF_EPOCH, CDF_EPOCH16, CDF_TIME_TT2000).

This is a change to the ValueInfo interface, i.e. the core of STIL.
Since ValueInfo is mostly used via the DefaultValueInfo implementation
however, the change is not likely to require source code changes
to client code.