ttools: remove range limitation on Subrange values.
Previously the (lo,hi) values of a subrange were constrained to be in
the range 0..1.  There is no particular reason for this limitation,
so it's been removed.  I thought I wanted to do something with longer
ranges, but at present I don't think I do, so this extended functionality
is not currently used or tested.