ttools: add new auto transparency shape mode.
This does transparent shading but adjusts the alpha channel automatically
according to how many pixels are being hit.  The details of how it does
this could probably be improved (it currently just looks at the single
most-filled bin and scales alpha based on that).

Also renamed and relabelled the existing (fixed) transparency mode,
which the more I think about the more I think is not a lot of use
in practice, at least interactively.