task: add option to ChoiceParameter so it can take class names.
With one of the constructors, a ChoiceParameter allows a classname
as a string value, if the named class is of the appropriate type
and it has a no-arg constructor.  This is a powerful pluggability
mechanism that I should implement across the board in the TASK
package, but it relies on generics and I don't have time to do
it passim for now.