ttools: add pluggable navigator framework for plot2 plots.
Instead of hard-coded per-Surface[Factory] mouse gesture handling for
navigation, navigation is now performed by pluggable Navigator objects
in the plot2 framework.  That means that in principle you can change
the navigation from the UI to, for instance, alter the zoom factor or
fix it so that only one rather than two axes zoom when the mouse
wheel is used.
So far this is just framework and not doing useful work, since the
UI does not provide mechanisms for adjusting the navigator.

Also cleared out the plot2/ZoomListener class because it's not used
any more.  Note however that this did attempt to map keyboard gestures
to zoom gestures as well, which is currently not in place.
However, this keyboard mapping didn't seem to work very well, for
reasons I think connected with keyboard focus not going/staying where
you want it.