topcat: fix table concatenation bug.
Bug reported when trying to concatenate two tables of which one has been
edited by double-clicking on a cell and replacing its content.
That replaces the column by an EditableColumn, which in turn means that
the ColumnInfo gets replaced in a somewhat inconsistent way in the
TopcatModel.  The subsequent matching between ColumnInfos required
during the concatenated table construction got the colinfos from different
places and hence failed.  Fix it so it gets them from the same places
for the test.  That fixes this error, though it's possible the same
thing is lurking elsewhere.