ttools: rewrite tick generation
Axis major tick labelling is now done in a more modular and extensible way,
and also in a way which is sensitive to the captioner and orientation.
This means that tick labels are generated in such a way that they will
not overlap, rather than generating them and weeding out overlapping ones
more or less at random.  In this way the visible ticks form a more
logical set.

This is done for linear, log and time axis labelling.