ttools: fix broken non-table-producing JyStilts commands.
Some commands (pixfoot and tcube) do not output tables but are
instances of ConsumerTask.  The JyStilts script was incorrectly
identifying these as commands that present a table as their
primary output.  This caused both a documentation bug, and an
execution bug, since the ProcessingMode that did the work for
them was not getting invoked.  Fix the identification in JyStilts
so that only tasks which output via ChoiceMode are marked thus;
others are executed without special arrangements for making the
result available in python.