registry: fix namespace in SOAP RI1.0 Search request
There was an error in the SOAP request sent for RI1.0 Search operations
by the registry client.  It was generating SOAP messages that started
like this:

   <rs:Search xmlns:rs='http://www.ivoa.net/wsdl/RegistrySearch/v1.0'>
     <rs:Where xmlns:rs="http://www.ivoa.net/xml/ADQL/v1.0"
               xmlns:ad="http://www.ivoa.net/xml/ADQL/v1.0"
               xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance">
       <ad:Condition xsi:type="ad:likePredType">

The remapping of the rs prefix to the ADQL namespace is clearly wrong.
This was spotted by Menelaus Perdikeas at ESAC.

The problem was related to the fact that the Where element is generated
(by the Where2DOM class in the RAYREG package) in the ADQL namespace,
but it needs to be in the RS namespace.  I had noticed that, and
attempted to patch it up with the line:

   whereEl.setPrefix( "rs" );

Unfortunately, this does something completely different from what
I intended/hoped/assumed.

I've now rewritten it using DOM methods to generate the right starting
elements, namely:

   <rs:Search xmlns:rs="http://www.ivoa.net/wsdl/RegistrySearch/v1.0">
     <rs:Where>
       <ad:Condition xmlns:ad="http://www.ivoa.net/xml/ADQL/v1.0"
                     xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
                     xsi:type="ad:likePredType">

The broken code was in place for many years.  Although it worked with
the AstroGrid and Euro-VO registries, and an earlier incarnation of
the STSci/NVO/VAO registry service, it did not work after about mid-2010
with the STSci one.