ttools: right-button/shift zoom drag option
When you drag the mouse in plane or time plots, you pan the plot.
Now, if you do the drag either with the right button or with the
shift key depressed, it does an X/Y zoom instead.  In most cases
this is much better than having to use the mouse wheel all the time.
Can't think why I didn't think of it before.