ttools: add drag zoom to other plot types
Using a modified drag gesture to zoom is now added to the remaining
plot types: sky and cube.  These are isotropic zooms, so don't confer
much advantage over the existing wheel zoom gestures, but it's worth
adding them since: (1) it's good if you don't have a mouse wheel and
(2) if people get used to doing this in other plot types it's intuitive
to have them throughout.