ttools: allow equality matching of tables in stilts
The ConsumerTask.createProducer method now results in a StarTable instance
which can be compared for equality to other instances obtained in the
same way.  In some cases (plot2) it can give major advantages if two
tables can be identified as equivalent.

It would be even better if the execution Environment could return the
same table instance for separately specified tables.  Maybe in future.