jniast: fix DefaultGrf/Plot to work headless
As spotted by Satoshi Eguchi (NAOJ), the Plot class cannot be used in
a headless environment because it constructs a JFrame instance.
The reason it does this is to be able to supply a suitably initialised
JComponent instance to DefaultGrf as required by the DefaultGrf constructor.

There are various ways around this, but it looks like DefaultGrf
doesn't really need a JComponent in any case, it just wants a
FontRenderContext.  This change gives DefaultGrf a new constructor
that only needs a FontRenderContext, which means that Plot no longer
needs to construct a JFrame.