ttools: allow plotter to have multiple positions
Replace Plotter's hasPosition method with getPositionCount.
This is effectively a generalisation so that a plotter can
plot tuple sequences which represent zero, one or more positions
in data space, rather than just zero or one.

TuplePlotter, as well as getting some relevant changes, is also
renamed to AbstractPlotter, which is a more appropriate name.

JELDataSpec is refactored slightly so that Coords and associated JEL
expressions are aggregated into a new class CoordValue.  This makes it
tidier anyway, but especially with the way that coordinates are now
processed in Plot2Task.