ttools: add pair plotter
Pair plotting is now done by a plotter with two positions, rather than
a single-position plotter with two extra non-positional coordinates.
This sidesteps the problem with changing DataGeoms, and generally
makes more sense.

Plotting is currently very basic (the only style option is colour).

Line drawing is done with a LineGlyph class, which is a placeholder
for a better implementation to come; it is currently grossly
inefficient and not thread-safe.