ttools: fix plan bug in LabelPlotter
A plan was being construed as matching even when the value type of the
associated cache map did not match (String instead of DepthString).
Actually, I'm not sure how this could happen, because I thought that
2D plots would always use String and 3D plots would always use DepthString,
but the fix, which involves storing Class<T> in the relevant places and
testing for equality, is harmless in any case.