ttools: plot 2D pair lines without both ends in bounds
Previously pair links were only drawn if the points at both ends
fell in the plot region.  Now for 2D the line will be drawn anyway,
whether both, either or neither of the end points are in the
plotting region.

It would be nice to get this in 3D as well, but it's harder.  Maybe one day.