ttools: plot 3D pair lines if only one end is in range
Pair lines are now plotted if either end is in the cube, it doesn't
require that both are.  This does mean that you get lines poking
out of the cube sometimes.  It would be better to truncate the lines
as they pass out of the cube faces.  Maybe one day.