ttools: generalise position to multi-position plots
A new plot form is added which lets you plot pairs of markers rather than
single points.  To support this, the MarkForm class is generalised to
provide implementations with 1 or more points plotted using the same marker.

Alongside these changes, all access to data positions using DataGeoms
now takes account of the fact that there may be any number
(not just zero or one) positions in a tuple.
This required changes to PointCloud.  The resulting PointCloud/SubCloud
class is working fine, but the interface is a bit messy, and should be
tidied up later.