ttools: taplint now checks for the correct ObsCore ID URI
The taplint OBS (ObsCore) validation stage checks whether the capabilities
reports implementation of the ObsCore DM before attempting ObsCore-specific
tests.  The old check unfortunately tested for the wrong string:
"ivo://ivoa.net/std/ObsCore-1.0" instead of
"ivo://ivoa.net/std/ObsCore/v1.0".
This error was propagated from example text in the TAPRegExt 1.0 standard
document.

Now, it checks for both strings, and reports a warning if it finds the
wrong one, though it takes either, or in fact anything with the
(case-insensitive) string "obscore" in it, as license to do the ObsCore
tests.

It also now tests the ObsCore ID case-insensitively, since Markus points
out that IVOA Identifiers 1.12 says all IVO IDs are to be considered
as case-insensitive.