ttools: make certain ShapeForm implementations effectively @Equality
SizeForm and PairLinkForm are now singletons.  MarkForm overrides
equals and hashCode.  This is so that certain distinct but equivalent
ShapePlotters can be identified as the same, though neither MarkForm
nor Plotter is currently declared with the annotation @Equality.
This is used by MatchPlotter so that adding layers can be done by
inserting layer attributes into existing layers, so it makes it
tidier rather than being essential.  Maybe declare MarkForm as
@Equality at some point, it could be done.