topcat: add plot result button to pair match plots
You can now hit a button when a pair match is completed and see the
input catalogues and output links.

So far it only works for pair matches, mainly because the pair link
information is not present in the result tables for single- and multi-table
matches, which means it's impossible(?) to get set up the GUI to do
the right plots.  It would be possible to do the right plots
programmatically using the calculated links - maybe do this in the future.