ttools: abstract plotter coord arrangement to new CoordGroup class
The new CoordGroup class now encapsulates all the information about
how the coordinates in a Plotter are arranged and how to find them
in a corresponding DataSpec ginve a corresponding DataGeom.
It also contains some additional information like which coordinates
should trigger a re-range if they are changed.

This refactoring is mainly because the information I require about
coordinate organisation is somewhat miscellaneous, and it's not
clear what is the right API for this functionality.  By abstracting
it out to a class separate from Plotter I can minimise the disruption
required by future changes to the interface.  The particular improvements
this time round is the addition of getRangeCoordIndices, and
tidying up the question of what is a partial coordinate, previously
determined in a hacky fashion, now explicitly stated by the CoordGroup.

The refactoring is not quite complete.  At this commit, ShapePlotter
and some of the related classes still have coordinate DataSpec indices
hard-coded.