topcat: histogram has a Bars control alongside Axes control
Refactored so that the AxisControl class is replaced by an AxisController
class that owns one or more Controls.  This allows the histogram plot
to control a Bars control alongside the Axes control.  That looks easier
to understand in the GUI.  For now, all the other AxisController
concrete subclasses only have a single control.