fits: improve nullable integer logic
Avoid a pre-processing pass to check whether there are nullable integers
in each column if all such columns are declared with all the metadata they
need in relation to nullability (i.e. a blank value).
In this way, by declaring a blank value for nullable columns you can
avoid a first pass for FITS output.