topcat: handle point clouds differently for counting
Point counting was not working properly, because a count was made for
each point SubCloud, and you could have one of those for each subset.

So introduce a new class TableCloud which contains all the points with
the same position coordinates, but aggregates those with different
subsets.  In this way each position in the point cloud is only
iterated over once.  This fixes errors in the displayed point count,
and also means that other operations that need to iterate over
displayed positions (e.g. blob inclusion) can work more efficiently.