topcat: get fromVisible selection working for partial positions
Histograms and Spectrograms have "partial positions" - their data
points have NaN values on at least one axis.  It still makes sense
to allow selection of the points from the visible range of the plot.
Fix it so that this can be done by handling partial positions
explicitly in the fromVisible scans, which is now separate from,
but shares code with, the blob shape scans.