ttools: add 2MASS PSC raw ASCII file input handler to examples
The examples sub-package now contains all the classes required to
read data from the DVD/FTP version of the 2MASS Point Source Catalogue.
Previously the PostgresAsciiStarTable class was in the formats
subpackage, but the additional information to actually use it was
not in the repository.  This can act as a template for other custom
large text file readers.