topcat: adjust plot interruption logic
When a UI action triggers a new plot, it's hard to work out whether to
cancel one that is currently in progress.  I noticed a regression issue;
the current implementation is less responsive when dragging sliders to
change the density scales etc.  Try to fix this up (success) without
making other things work less well (?).
It's hard to work out whether to cancel a previous plot when starting a
new one (triggered by a UI action).