ttools: add hooks for specifying pixel compositor
The Compositor is no longer hard coded into the PaperTypeSelector,
clients of that class must supply their own.  Plot2Task provides a
parameter for specifying this (saturation or a specified boost).
Currently topcat does not, it still uses a hardwired value,
but the hook is there for changing it in the future if required.