topcat: fix shader control so it doesn't disappear too much
The shader control is only visible in the stack if one of the plotters
in the stack is capable of using it (producing layers with aux shading).
However, if this is interpreted too literally it can disappear in an
annoying way, e.g. if you are just hiding one of the layers.
Fix it so that the shader control doesn't disappear unless there are
zero aux-capable layer controls in the stack, rather than just zeor
active aux-capable layer controls in the stack.