topcat: improve sky plot re-ranging
Refactored the auto-ranging logic so it's mostly in the AxisController
superclass rather than the subclasses.  Not sure why it was separated
before, hopefully there wasn't a good reason I'm missing.
Now the code is tidier plotting a new dataset on a sky plot will
re-range the axes as it does for a plane plot.