ttools: navigation gestures now yield NavAction not aspect objects
This change adds a hook which allows plot navigation gestures to return
a GUI decoration with a visual indication of the navigation action
that is in progress as well as the instruction to change the surface
view.  This may make it more obvious from the GUI how to drive the
navigation.  At present, only the hooks are present, no actual GUI
decorations are implemented.