topcat: move plot2 stack control toolbar buttons
The buttons for adding and removing controls to the control stack in
plot2 windows have been moved out of the main toolbar and into a
new sub-toolbar just above the control stack.  This is a more logical
and intuitive place to find them, especially when the control panel
is floated away from the plot display itself.  Secondary considerations:
it uses up more vertical pixels, which are at a premium especially
on laptop screens (bad), but it also gives more room for expansion
for more different layer types, which hopefully will be added in the
future (good).