ttools: add option to anchor plane X/Y axes at data value of zero
This is specially useful for histogram, since it means you can fix
the X axis to sit at the bottom of the plot.
Realised while doing this that the configuration parameters for
navigators really belong with the SurfaceFactory not with the Navigator
itself.  Rectified this for PlaneSurfaceFactory (and histogram),
but not for Time, Sky, Cube etc.