topcat: adjust ControlWindow layout
Fix the window size so that it's determined by its contents and the
size is not hard-wired - although it looked OK on my machine, on Mac
the window was a bit small and needed scrolling.

Also add a few pixels spacing where some of the components were squashed
together and looked a bit ugly.