ttools: add antialiasing option for plot grid text/line drawing
A new config option, GRID_ANTIALIAS, is added for all plot types.
This is partly so you can choose whether you want grids and labels
antialiased.  However, it's also to work around a MacOS java bug
that paints non-horizontal text all wrong.  If the text is antialiased,
it gets it right.  So on MacOS default to antialiased.  This isn't
particularly satisfactory (antialiased is slower, and I don't like
plaform-dependent behaviour) but it's better than having backwards text.