topcat: subscribe to new experimental table.get.stil SAMP MType
This allows external clients to get a URL for one of the tables currently
loaded into topcat.  This functionality was suggested by @taldcroft
in the discussion https://github.com/astropy/astropy/pull/1920.

The MType (table.get.stil) resembles the existing table.load.stil.
Possibly there should be a table.get.votable etc, but that should
probably be decided with consultation from the SAMP community.
The machinery is present to add those - it would be a one-liner.