ttools: abstract mouse button number handling
All determination of which mouse button is in use (changed or pressed)
now goes via the PlotUtil.getButton*Index methods rather than examining
the MouseEvent directly.  The point of this is to have control over
what mouse/keyboard gestures count as logical button 1/2/3.
It's not straightforward to work out what should mean what across
different OSes and mouse devices, doing it like this means we actually
know what's going on, and can document it at the user level accordingly.