ttools: refactor captioner configuration
The logic for selecting a font from configuration options is now
broken out into a new class CaptionerKeySet rather than using
opaque logic within the StyleKeys class.  This means it can be
reused, which in turn means the LabelPlotter class is able to use
a different instance of the relevant config keys than that used
for the axis labelling.  So now the logic in the topcat GUI classes
lets you choose a different font for point labelling than for
axis labelling.