ttools: mirror quadrants for band zoom actions
The Frame Zoom actions allow you to drag out a frame to focus in on
(or do something similar for zooming out).  However the sense of
these presents a problem: the sensible way to have zoom in on the Y
axis in the up direction, but behaviour of the corresponding function
in the old plot windows had zoom in going down (this older behaviour
came from PtPlot, which was presumably following AWT graphics coordinate
sense).

Solve this problem by allowing either up or down to be a zoom in (or out);
what determines the in/outness is just the X component of the drag
(right is in, left is out).  This comes down to mirroring the drag
actions in the previously unused quadrants (up-right, down-left).