topcat: preserve plot form config options when table is changed
Previously the global config options for Forms were reset when the
table was changed in a plot.  This was annoying if you were changing
a table but wanted to keep basically the same plot (reported by
Phil Evans, Leicester).  Fix it so that the old options are copied
over to the new style panel.