vo: update ADQL library to v1.2
Gregory Mantelet's ADQL library is updated to version 1.2, which
fixes a number of parsing errors.

URLs for the original library are at:
   https://github.com/gmantele/taplib
   http://cdsportal.u-strasbg.fr/adqltuto/

However I had to make some adjustments and rebuild, mostly to make it
work at java 1.5 (the distributed version is for java 7).
The version I used is at https://github.com/mbtaylor/taplib on branch java5,
for this commit the taplib SHA1 d5d11158a301c04a8d8.