ttools: rework taplint XSD validation
Some of the taplint stages (CPV, AVV, TMV) validate TAP service documents
against standard IVOA XML Schemas.

The way this was done previously was to prepare an XSD validator that
knew about the XSD schema expected for the given document.  Unfortunately,
it seems to be impossible to combine that approach with allowing the
document to specify additional schemas (for additional namespaces in
the same document).  That capability was required for the GACS TAP
service, that is using a custom size attribute in a foreign namespace
(permitted by the schema).  To accommodate that, I had to rework the
validation framework.  It now gets schemas as declared by the client
documents, rather than hard-wiring them into the validator, but uses
a custom schema resource resolver to suppy official versions of the
expected schemas, and reports on whether the expected
schemas/namespaces/element names were used.

It seems remarkably painful to get all this working, and it's possible
there are still some mistakes.

However, it now works better than before in some respects, in particular
the CPV stage (for which the schema was absent from its expected home
on the IVOA website) works rather than not working, and so is now
turned on (rather than disabled) by default.