ttools: add new png-transp graphics output format
This is just like PNG format but the background is transparent.
PNG output used to be transparent, but I changed it before v4
because it was causing problems for some users - specifically
I think the background is transparent black not transparent white.
So default PNGs are still opaque, but you can have transparent
if you want to.