ttools: rationalise RA/Dec parameters
Add a new class SkyCoordParameter for use whenever task parameters
representing RA and Dec are required.  This reduces typing and
increases consistency by putting the documentation in one place.

Also include in that class utilities for guessing RA and Dec values
for use with a table.  That functionality was previously only available
for tskymatch2, but it makes just as much sense for other RA/Dec-based
tasks like pixfoot and the various sky matching commands (coneskymatch,
sqlskymatch, cdsskymatch).  Those are now therefore easier to use,
you don't need to type in RA and Dec column names.