ttools: generalise upload match FindMode
Add some more options for the upload find mode, corresponding to things
that you can do to a table in-place (possible if the output rows match
1:1 with the input rows).

This involves restructuring the way that the FindModes are done;
separate the service-directed modes from the user-directed ones.
The user-directed ones are specific to the UI, so the ttools one
will differ from the topcat one.