vo: make TAP load window jobs window slightly less annoying
The "Running Jobs" window previously scrolled the Job Details panel
to the bottom of the scroll pane every time a new job was selected
or the job details were refreshed.  It did this in a jerky way.
I've stopped the jerkiness and made it scroll to the top instead,
which is an improvement, but I don't know how to stop it changing
the scroll position at all.

Also moved the field for Job URL display outside the scroll pane
so it never gets scrolled out of view.