ttools: add getIndex method to ConeQueryRowSequence interface
A ConeQueryRowSequence now has a record of the row index relative
to the underlying table (whatever that is).  This enables the use
of filter CQRSs that can modify the sequence of rows (for instance
skipping some) without losing the information that allows you to
stitch the results back together in the right order.