topcat: do not disable Activation options when no columns present
Previously, the "View URL As .." options in the ActivationQueryWindow
were disabled when there were no columns in the column selector
suitable for specifying URLs.  This was broken, because adding
a suitable (string) column did not re-enable it.  Moreover, since
the relevant selectors are editable, you can input a string expression
based on existing columns.  So remove this cleverness.

Reported by Igor Chilingarian.