ttools: use MOCs to pre-filter rows for upload xmatch
MOCs are retrieved from the CDS footprint service, and used in the
stilts cdsskymatch command to filter out any rows which are outside
the footprint of the service.

Minor change to a couple of bits of MOC reading implementation and
parameter usage of SkyConeMatch2 (nside is now null by default).
This is more logical, I think I did it different before because
I couldn't have null IntegerParameters.