topcat: cds upload match window now provides known aliases
Add a new class Downloader that manages downloading an object from
an external source (or in general performing fallible acquisition of
a resource which may take longer than is acceptable on the
Event Dispatch Thread).  Use this to retrieve the current list of
known VizieR table aliases from CDS, and display those in a selector
for the user to choose from.  The user can also type in full VizieR
table names (like II/246/out); although CDS does offer a download of
the full list of these, there are too many of them to offer
multiple choice.