vo: add progress monitor for registry endpoint updater
When the list of registries is queried to update the list of available
registries, Downloader monitor component is displayed to show progress.

This is useful, and an improvement, but it should really work
differently than this - there ought to be a single registry selection
model for the whole application not one for each window.  One day.