ttools: make it possible to interrogate plot component
Adjusted the API so that if you get a JComponent from Plot2Task it
is explicitly typed as a PlotDisplay, and PlotDisplay has some methods
on it to determine the current plot surface.  This means that you
can (for instance) programmatically examine a plot you created,
and which may have been interacted with by a user, to determine things
like its current field of view.