ttools: fix it so upload matches don't upload zero-row blocks
Add some more logic to BlockUploader so that it's possible to determine
whether a block of queries is empty before sending it to the
UploadMatcher.  This is a bit fiddly, but worth doing to avoid uploading
an empty block in the case that an exact number of blocks is required.
Requested/suggested by Thomas Boch, since efficiency aside there is
currently a bug in the CDS upload matcher service meaning empty uploads
generate an error.